fix(shipping): resolve getShipmentsRepository errors by moving logic to ShipmentRepository
