+++ back/docker-compose.yml
@@ -45,7 +45,7 @@ services:
     env_file:
       - .env
     ports:
-      - "3010:3010"
+      - "3000:3000"
     networks:
       - shipping_network
