+++ back/src/shipping/repositories/mysql/mysql_shipment.repository.ts
@@ -22,4 +22,25 @@ export default class MySqlShipmentRepository implements ShipmentRepository {
             relations: ['user', 'transportMethod', 'originAddress', 'destinationAddress', 'shipmentProducts', 'shipmentProducts.product']
         });
     }
+    
+    async findAll(page: number, itemsPerPage: number): Promise<[Shipment[], number]> {
+        const skip = (page - 1) * itemsPerPage;
+
+        return await this.shipmentRepository.findAndCount({
+            relations: [
+                'user',
+                'originAddress',
+                'destinationAddress',
+                'transportMethod',
+                'shipmentProducts',
+                'shipmentProducts.product',
+                'logs'
+            ],
+            skip,
+            take: itemsPerPage,
+            order: {
+                createdAt: 'DESC',
+            },
+        });
+    }
 }
