+++ back/src/shipping/shipping.module.ts
@@ -13,8 +13,6 @@ import TransportMethodsRepository from './repositories/transport_methods.reposit
 import MySqlTransportMethodsRepositories from './repositories/mysql/mysql_transport_methods_repositories';
 import ShipmentRepository from './repositories/shipment.repository';
 import MySqlShipmentRepository from './repositories/mysql/mysql_shipment.repository';
-import GetShipmentsRepository from './repositories/get-shipments.repository';
-import { MysqlGetShipmentsRepository } from './repositories/mysql/mysql_get_shipments.repository';
 import { ShippingLog } from './entities/shipping-log.entity';
 
 @Module({
@@ -40,10 +38,6 @@ import { ShippingLog } from './entities/shipping-log.entity';
     {
       provide: ShipmentRepository,
       useClass: MySqlShipmentRepository
-    },
-    {
-      provide: GetShipmentsRepository,
-      useClass: MysqlGetShipmentsRepository
     }
   ]
 })
